feat: support custom report dataformat like Taskwarrior for `due` datetimes (#636)
To support taskwarrior's`report._.dateformat` option, this PR reads this configuration parameter
and uses it when formatting the `due` task property.

The only slight complexity in implementation is converting from
Taskwarrior's date formatting to `chrono` date format. I want to spend a
little more time cleaning this up, but wanted to post this draft PR to
see if this is an acceptable feature to add.